weekend

+++ src/main/webapp/assets/js/page.js
@@ -60,13 +60,19 @@ var Page = {
 		// 初始化表头
 		var thead = $("<thead></thead>");
 		var theadTr = $("<tr></tr>");
+		//checkbox
+		theadTr.append($("<th></th>").append($("<input type='checkbox' onclick='javascript:Page.checkAllItems(this);'>")));
+		
 		$.each(cols, function(i, item) {
 			var sortStr = item.noSort ? ""
 					: "class='sorting' onclick='Page.sorting(this)'";
 			theadTr.append($("<th " + sortStr + "></th>").attr("cName",
 					item.cName).html(item.cValue));
 		});
-		theadTr.append($("<th class='span2'></th>").html("操作"));
+		//判断菜单数组长度如果不为空才添加操作列
+		if(typeof(perms) != "undefined" && perms.length >0 ){
+			theadTr.append($("<th class='span2'></th>").html("操作"));
+		}
 		thead.append(theadTr);
 		tBody = $("<tbody></tbody>");
 		$(obj.tableId).empty().append(thead).append(tBody);
@@ -88,6 +94,8 @@ var Page = {
 		var tbody = $(Page.defaultVal.tableId + " tbody:first").empty();
 		$.each(pageData.data, function(i, item) {
 			var tr = $("<tr></tr>");
+			//checkbox
+			tr.append($("<td></td>").append($("<input type='checkbox' data-id="+item.id+">")));
 			$.each(Page.defaultCols, function(j, col) {
 				if (col.format) {
 					tr.append($("<td></td>").html(
@@ -96,9 +104,13 @@ var Page = {
 					tr.append($("<td></td>").html(item[col.cName]));
 				}
 			});
-			tr.append($("<td></td>").html(Page.operBtn(item.id)));
+			if(typeof(perms) != "undefined" && perms.length >0 ){
+				tr.append($("<td></td>").html(Page.operBtn(item.id)));
+			}
 			tbody.append(tr);
 		});
+		
+		Page.registerTdOnclickEvent();
 	},
 
 	/**
@@ -164,25 +176,35 @@ var Page = {
 	},
 	// 生成操作按钮
 	operBtn : function(id) {
-		return App.initDropDownBtn({
-			icon : "icon-cogs",
-			name : "操作",
-			group : [ {
+		var group = [];
+		if(App.in_array("update",perms)){
+			group.push({
 				clickFn : "Page.updateObj(" + id + ")",
 				name : "修改",
 				icon : "icon-pencil"
-			}, {
+			})
+		}
+		if(App.in_array("delete",perms)){
+			group.push({
 				clickFn : "Page.deleteObj(" + id + ")",
 				name : "删除",
 				icon : "icon-trash"
-			}, {
+			})
+		}
+		if(App.in_array("show",perms)){
+			group.push({
 				clickFn : "Page.viewObj(" + id + ")",
 				name : "查看",
 				icon : "icon-search"
-			} ]
+			})
+		}
+		return App.initDropDownBtn({
+			icon : "icon-cogs",
+			name : "操作",
+			group: group
 		});
 	},
-	updateObj : function(id) {
+	/*updateObj : function(id) {
 		window.location.href = Page.subUrl() + "/update/" + id;
 	},
 	deleteObj : function(id) {
@@ -190,6 +212,57 @@ var Page = {
 	},
 	viewObj : function(id) {
 		window.location.href = Page.subUrl() + "/show/" + id;
+	},*/
+	
+	updateObj : function() {
+		var boxes = $(Page.defaultVal.tableId).find("td :checkbox:checked");
+		if(boxes.length == 0){
+			App.alert("请选中一条记录");
+			return;
+		}
+		if(boxes.length > 1){
+			App.alert("请只选择一条记录");
+			return;
+		}
+		var id = boxes.first().attr("data-id");
+		window.location.href = Page.subUrl() + "/update/" + id;
+	},
+	
+	viewObj : function(id) {var boxes = $(Page.defaultVal.tableId).find("td :checkbox:checked");
+		if(boxes.length == 0){
+			App.alert("请选中一条记录");
+			return;
+		}
+		if(boxes.length > 1){
+			App.alert("请只选择一条记录");
+			return;
+		}
+		var id = boxes.first().attr("data-id");
+		window.location.href = Page.subUrl() + "/show/" + id;
+	},
+	
+	deleteObj : function(id) {
+		var boxes = $(Page.defaultVal.tableId).find("td :checkbox:checked");
+		if(boxes.length == 0){
+			App.alert("请选中一条记录");
+			return;
+		}
+		var delFun = function(result){
+			if(!result) return;
+			var f = document.createElement("form");
+			document.body.appendChild(f);
+			$.each(boxes,function(i,item){
+				var i = document.createElement("input");
+				i.type = "hidden";
+				i.name = "ids";
+				i.value = $(item).attr("data-id");
+				f.appendChild(i);
+			});
+			f.action = Page.subUrl() + "/delete";
+			f.method = "POST";
+			f.submit();
+		}
+		App.confirm(delFun);
 	},
 	/**
 	 * 截取url供修改，删除，查看等使用
@@ -235,8 +308,35 @@ var Page = {
 		Page.cqData.pageNo = 1;
 		Page.accToSer(Page.pageChangedCallBack);
 		$('#Pagination').bootstrapPaginator("showFirst");
+	},
+	
+	/**********************************************************
+	 * 改变方式把操作按钮都放到上边
+	 **********************************************************/
+	checkAllItems : function(obj){
+		if($(obj).attr("checked") == "checked"){
+			$(obj).closest("table").find(":checkbox").attr("checked","checked");
+		}else{
+			$(obj).closest("table").find(":checkbox").removeAttr("checked");
+		}
+	},
+	
+	/**
+	 * 给所有td注册事件，点击选中当前行tr的checkbox
+	 * @param $
+	 */
+	registerTdOnclickEvent : function(){
+		$("td").click(function(){
+			var box = $(this).parent().find(":checkbox");
+			if(box.attr("checked") == "checked"){
+				box.removeAttr("checked");
+			}else{
+				box.attr("checked","checked");
+			}
+		})
 	}
 };
+
 
 /*******************************************************************************
  * 分页封装插件 基于jquery pagination
